Add useFormatMemorySize hook
