+++ app/ui-admin/client/components/info/formatters.js
@@ -3,27 +3,6 @@ import s from 'underscore.string';
 
 export const formatNumber = (number) => s.numberFormat(number, 2);
 
-export const formatMemorySize = (memorySize) => {
-	if (typeof memorySize !== 'number') {
-		return null;
-	}
-
-	const units = ['bytes', 'kB', 'MB', 'GB'];
-
-	let order;
-	for (order = 0; order < units.length - 1; ++order) {
-		const upperLimit = Math.pow(1024, order + 1);
-
-		if (memorySize < upperLimit) {
-			break;
-		}
-	}
-
-	const divider = Math.pow(1024, order);
-	const decimalDigits = order === 0 ? 0 : 2;
-	return `${ s.numberFormat(memorySize / divider, decimalDigits) } ${ units[order] }`;
-};
-
 export const formatDate = (date) => {
 	if (!date) {
 		return null;
